refactor(android): move mrzqrscanner packages
