+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.proofofpassportapp.mrzqrscanner">
+    package="com.selfxyz.mrzqrscanner">
 
     <application>
         <activity
-            android:name="com.proofofpassportapp.PhotoPickerActivity"
+            android:name="com.selfxyz.mrzqrscanner.PhotoPickerActivity"
             android:exported="false"
             android:theme="@style/Theme.AppCompat.Translucent" />
     </application>
