+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/MRZQRScannerPackage.java
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BUSL-1.1; Copyright (c) 2025 Social Connect Labs, Inc.; Licensed under BUSL-1.1 (see LICENSE); Apache-2.0 from 2029-06-11
 
-package com.proofofpassportapp;
+package com.selfxyz.mrzqrscanner;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.NativeModule;
@@ -13,8 +13,8 @@ public class MRZQRScannerPackage implements ReactPackage {
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
         return List.of(
-                new com.proofofpassportapp.ui.PassportOCRViewManager(reactContext),
-                new com.proofofpassportapp.ui.QRCodeScannerViewManager(reactContext)
+                new com.selfxyz.mrzqrscanner.ui.PassportOCRViewManager(reactContext),
+                new com.selfxyz.mrzqrscanner.ui.QRCodeScannerViewManager(reactContext)
         );
     }
