+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/PhotoPickerActivity.java
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: BUSL-1.1
 // NOTE: Converts to Apache-2.0 on 2029-06-11 per LICENSE.
 
-package com.proofofpassportapp;
+package com.selfxyz.mrzqrscanner;
 
 import android.content.Intent;
 import android.net.Uri;
