+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/fragments/CameraFragment.kt
@@ -1,4 +1,4 @@
-package com.proofofpassportapp.fragments
+package com.selfxyz.mrzqrscanner.fragments
 
 import android.Manifest
 import android.app.AlertDialog
