+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/mlkit/FrameMetadata.kt
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.proofofpassportapp.mlkit
+package com.selfxyz.mrzqrscanner.mlkit
 
 /** Describing a frame info.  */
 class FrameMetadata private constructor(val width: Int, val height: Int, val rotation: Int, val cameraFacing: Int) {
