+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/ui/QRCodeScannerViewManager.kt
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BUSL-1.1; Copyright (c) 2025 Social Connect Labs, Inc.; Licensed under BUSL-1.1 (see LICENSE); Apache-2.0 from 2029-06-11
 
-package com.proofofpassportapp.ui
+package com.selfxyz.mrzqrscanner.ui
 
 import android.view.Choreographer
 import android.view.View
