+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/utils/MRZUtil.kt
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: BUSL-1.1
 // NOTE: Converts to Apache-2.0 on 2029-06-11 per LICENSE.
 
-package com.proofofpassportapp.utils
+package com.selfxyz.mrzqrscanner.utils
 
 
 import org.jmrtd.lds.icao.MRZInfo
