+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/utils/OcrUtils.kt
@@ -1,4 +1,4 @@
-package com.proofofpassportapp.utils
+package com.selfxyz.mrzqrscanner.utils
 
 import android.util.Log
 import com.google.mlkit.vision.text.Text
