+++ app/android/android-mrz-qr-scanner/mrzqrscanner/src/main/java/com/selfxyz/mrzqrscanner/utils/QrCodeDetectorProcessor.kt
@@ -13,7 +13,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package com.proofofpassportapp.utils
+package com.selfxyz.mrzqrscanner.utils
 
 
 import android.graphics.Bitmap
@@ -26,8 +26,8 @@ import com.google.zxing.RGBLuminanceSource
 import com.google.zxing.Result
 import com.google.zxing.common.HybridBinarizer
 import com.google.zxing.qrcode.QRCodeReader
-import com.proofofpassportapp.mlkit.FrameMetadata
-import com.proofofpassportapp.utils.ImageUtil
+import com.selfxyz.mrzqrscanner.mlkit.FrameMetadata
+import com.selfxyz.mrzqrscanner.utils.ImageUtil
 import io.fotoapparat.preview.Frame
 import java.nio.ByteBuffer
 import java.util.concurrent.ExecutorService
